sdfsdfdsf

+++ src/database/migrations/2023_12_03_025429_add_column_visited_posts_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('posts', function (Blueprint $table) {
-            $table->mediumInteger('visited')->default(0)->nullable();
+            $table->smallInteger('visited')->default(0)->nullable();
 
         });
     }
